Worked to re-align frontend, backend and common

+++ web/web_common/src/api/oauth/providers.rs
@@ -1,13 +1,14 @@
 //! Struct defining an OAuth2 login provider supported by the backend.
-use crate::combine_path;
 use serde::{Deserialize, Serialize};
+
+use crate::combine_path;
 
 pub const ENDPOINT: &str = "/providers";
 pub const FULL_ENDPOINT: &str = combine_path!(super::FULL_ENDPOINT, ENDPOINT);
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub struct OAuth2LoginProvider {
-    pub id: i32,
+    pub id: i16,
     pub name: String,
     pub icon: String,
     pub client_id: String,
